+++ app/controllers/app.js
@@ -723,7 +723,7 @@ function getContentType(format){
 
 function setCrossDomain(res){
     res.header("Access-Control-Allow-Origin", "*");
-    res.header("Access-Control-Allow-Headers", "X-Requested-With");
+    res.header("Access-Control-Allow-Headers", "X-Requested-With, X-Prototype-Version, X-CSRF-Token");
 }
 
 function generateCacheKey(database,tables,is_authenticated){